pyproject.toml
[project]
name = "pulse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorlog>=6.10.1",
    "pydantic>=2.12.5",
    "psycopg2-binary>=2.9.11",
    "pre-commit>=4.5.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
override-dependencies = ["psycopg2>=2.9"]

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
pipeline = [
    "click>=8.3.1",
    "dbt-postgres>=1.8",
    "dlt[postgres]>=1.24.0",
]
web_app = [
    "alembic>=1.18.4",
    "flask>=3.1.3",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=25.3.0",
    "pyyaml>=6.0.3",
]
development = [
    "dotenv>=0.9.9",
    "mypy>=1.15.0,<1.20.0",
    "pytest>=9.0.2",
    "ruff>=0.15.8",
]

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py312"
exclude = ["dbt-tmc"]

[tool.ruff.lint]
extend-ignore = [
    "F401",
    "E203",
    "E266",
    "E402",
    "E501",
    "E711",
    "E712",
    "F403",
    "F821",
]
select = ["E", "F", "I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.mypy]
# Strict type checking configuration
strict = true
ignore_missing_imports = true
mypy_path = "."
explicit_package_bases = true
disable_error_code = ["no-redef"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
filterwarnings = ["ignore::FutureWarning"]
